requirements.txt
flask
gunicorn
pdf2docx
pypdf
pytesseract
pdf2image
Pillow
